(fix) add transacting calls to getDrivers and updateDrivers
